+++ DI-App/Controllers/CalendarVC/CalendarVC.xib
@@ -57,11 +57,10 @@
                     <items>
                         <navigationItem id="t4l-tC-taB">
                             <barButtonItem key="leftBarButtonItem" title="Item" image="hamburger-icon" landscapeImage="Image" largeContentSizeImage="hamburger-icon" springLoaded="YES" id="aFD-n9-JEc">
-                                <inset key="imageInsets" minX="-50" minY="0.0" maxX="0.0" maxY="0.0"/>
                                 <inset key="largeContentSizeImageInsets" minX="0.0" minY="60" maxX="0.0" maxY="0.0"/>
                                 <color key="tintColor" red="1" green="0.14913141730000001" blue="0.0" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
                                 <connections>
-                                    <action selector="menu:" destination="-1" id="50z-Lc-K1O"/>
+                                    <action selector="menu:" destination="-1" id="knG-9j-lvd"/>
                                 </connections>
                             </barButtonItem>
                             <barButtonItem key="rightBarButtonItem" systemItem="add" id="CfP-U1-hNL">
@@ -100,6 +99,6 @@
     </objects>
     <resources>
         <image name="Image" width="1200" height="1004"/>
-        <image name="hamburger-icon" width="128" height="128"/>
+        <image name="hamburger-icon" width="50" height="50"/>
     </resources>
 </document>
