+++ DI-App/Controllers/CalendarVC/SideMenuViewController.xib
@@ -16,9 +16,6 @@
             </connections>
         </placeholder>
         <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <viewController id="EOg-sF-35k" customClass="SideMenuViewController" customModule="DI_App" customModuleProvider="target">
-            <point key="canvasLocation" x="-401" y="-276"/>
-        </viewController>
         <view contentMode="scaleToFill" id="iN0-l3-epB">
             <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
             <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
