+++ DI-App/Controllers/ClassesVC/Classes.xib
@@ -12,8 +12,8 @@
     <objects>
         <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="Classes" customModule="DI_App" customModuleProvider="target">
             <connections>
+                <outlet property="mainView" destination="i5M-Pr-FkT" id="6Ok-32-65g"/>
                 <outlet property="tableView" destination="hpR-c1-gCm" id="GSC-aT-uWp"/>
-                <outlet property="view" destination="i5M-Pr-FkT" id="sfx-zR-JGt"/>
             </connections>
         </placeholder>
         <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
@@ -32,7 +32,14 @@
                 <navigationBar contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="A1O-1n-403">
                     <rect key="frame" x="0.0" y="20" width="375" height="44"/>
                     <items>
-                        <navigationItem title="Courses" id="Oiy-Qc-Ni6"/>
+                        <navigationItem title="Courses" id="Oiy-Qc-Ni6">
+                            <barButtonItem key="leftBarButtonItem" title="Item" image="hamburger-icon" landscapeImage="Image" largeContentSizeImage="hamburger-icon" id="d9Z-uV-EHk">
+                                <color key="tintColor" red="1" green="0.0" blue="0.0" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+                                <connections>
+                                    <action selector="menu:" destination="-1" id="m6z-vQ-vOf"/>
+                                </connections>
+                            </barButtonItem>
+                        </navigationItem>
                     </items>
                 </navigationBar>
             </subviews>
@@ -47,6 +54,11 @@
                 <constraint firstItem="hpR-c1-gCm" firstAttribute="top" secondItem="A1O-1n-403" secondAttribute="bottom" id="t2K-J3-gW6"/>
             </constraints>
             <viewLayoutGuide key="safeArea" id="fnl-2z-Ty3"/>
+            <point key="canvasLocation" x="58.5" y="134.5"/>
         </view>
     </objects>
+    <resources>
+        <image name="Image" width="1200" height="1004"/>
+        <image name="hamburger-icon" width="50" height="50"/>
+    </resources>
 </document>
