+++ DI-App/Controllers/Menu.swift
@@ -1,0 +1,16 @@
+//
+//  Menu.swift
+//  DI-App
+//
+//  Created by Adrien Meyer on 10/01/2018.
+//  Copyright © 2018 Developer.Institute. All rights reserved.
+//
+
+import UIKit
+import SideMenu
+
+class Menu: UIStoryboardSegue {
+    
+
+    
+}
